+++ Archive/ViewControllers/Login_Register/OnBoarding2VC.swift
@@ -28,8 +28,7 @@ class OnBoarding2VC : UIViewController {
     @objc private func goToNextViewController() {
         let nextVC = OnBoarding3VC() // 다음 보여줄 뷰컨트롤러로 변경해!
         navigationController?.pushViewController(nextVC, animated: true)
-//        nextVC.modalPresentationStyle = .fullScreen
-//               present(nextVC, animated: true, completion: nil)
+
     }
 
 }
