+++ Archive/ViewControllers/Login_Register/OnBoarding3VC.swift
@@ -35,8 +35,7 @@ class OnBoarding3VC : UIViewController {
     @objc private func goToNextViewController() {
         let nextVC = OnBoarding4VC() // 다음 보여줄 뷰컨트롤러로 변경해!
         navigationController?.pushViewController(nextVC, animated: true)
-//        nextVC.modalPresentationStyle = .fullScreen
-//               present(nextVC, animated: true, completion: nil)
+
     }
 }
